% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitulation_wrapper.R
\name{capitulation_wrapper}
\alias{capitulation_wrapper}
\title{A wrapper to use capitulation function with the GMM formalism}
\usage{
capitulation_wrapper(theta, ...)
}
\arguments{
\item{theta}{Structural parameters}

\item{...}{Additional parameters provided}
}
\description{
A wrapper to use capitulation function with the GMM formalism
}
