% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_moment.R, R/create_moment_data.R
\name{create_moment}
\alias{create_moment}
\alias{create_moment_data}
\title{Create a dataframe storing moments regarding wealth distribution}
\usage{
create_moment(
  EP_2015,
  EP_2018 = NULL,
  EP_lon = NULL,
  by = NULL,
  N_moments = 1,
  ages_cross_section = c(20, 80),
  ages = c(30, 65),
  stats = c("mean", "sd"),
  quantiles = c(0.1, 0.5, 0.9),
  scale_variable_moment1 = c("level", "log", "asinh"),
  scale_variable_moment2 = c("level", "log", "asinh"),
  scale_moment1 = c("level", "log", "asinh"),
  moment1 = c("level", "share"),
  stat_moment2 = c("proportion", "difference", "growth"),
  wealth_var = "PATFISOM",
  age_var = "AGEPR",
  deltaK_var = "dW",
  age_2015_var = "AGEPR_2015",
  moment_var = "moment_data",
  normalize = FALSE,
  survey_years = c(2015, 2018),
  exclude_negative = FALSE
)

create_moment_data(
  EP_2015,
  EP_2018 = NULL,
  EP_lon,
  data_microsimulated,
  observed_moment_data = NULL,
  probability_survival_var = NULL,
  r = 0.03,
  gamma = 0.5,
  beta = 0.98,
  r_low = NULL,
  r_high = NULL,
  r.parameters = NULL,
  gamma.parameters = NULL,
  beta.parameters = NULL,
  non_ricardian = FALSE,
  non_ricardian_var = NULL,
  N_moments = 3,
  ages = c(30, 65),
  stats = c("mean", "sd"),
  quantiles = c(0.1, 0.5, 0.9),
  age_var_simulations = "age",
  by = NULL,
  normalize = TRUE,
  scale_model = c("level", "log", "loglog"),
  scale_variable_moment1 = c("level", "log", "asinh"),
  scale_moment1 = c("level", "log", "asinh"),
  moment1 = c("level", "share"),
  scale_variable_moment2 = c("level", "log", "asinh"),
  stat_moment2 = c("proportion", "difference", "growth"),
  wealth_var = "PATFISOM",
  exclude_negative = FALSE,
  Hgiven_var = "H_given",
  Hreceived_var = "H_received",
  additional_vars = NULL
)
}
\arguments{
\item{EP_2015}{Wealth survey used for its cross-sectional dimension
or microsimulated population}

\item{EP_2018}{(Optional). Wealth survey used for its cross-sectional dimension
or microsimulated population}

\item{EP_lon}{Longitudinal wealth survey used for its time dimension}

\item{N_moments}{Number of moments needed. See \code{details} section 
for the ordered list of moments}

\item{ages}{Ages considered to restrict sample}

\item{stats}{Moment that should be used. This must be a
data summary statistics that should match a base R function}

\item{quantiles}{Quantiles that should be computed}

\item{wealth_var}{Wealth variable in survey}

\item{age_var}{Age variable in survey}

\item{deltaK_var}{Name for the growth variable that is temporarily created}

\item{age_2015_var}{Age variable storing age at first occurrence}

\item{moment_var}{Variable name for moment result}

\item{normalize}{Logical value indicating whether we want to normalize
variable}

\item{survey_years}{Collection years for the two survey waves}

\item{exclude_negative}{Logical value indicating whether we should
exclude negative wealth values}

\item{data_microsimulated}{Microsimulated data that follows \link[capitulation]{life_cycle_model}
expected structured}

\item{r}{Interest rate vector. Because we assume an individual
faces the same interest rate every year, only the first value
is used}

\item{gamma}{Risk-aversion coefficient}

\item{beta}{Discount factor}

\item{r_low}{Low interest rate if heterogeneity is introduced}

\item{r_high}{High interest rate if heterogeneity is introduced}

\item{r.parameters}{Parameters value that should be used if
r is a formula}

\item{gamma.parameters}{Parameters value that should be used if
gamma is a formula}

\item{beta.parameters}{Parameters value that should be used if
beta is a formula}

\item{non_ricardian}{Should we introduce some non-Ricardian household}

\item{age_var_simulations}{Age variable in simulated data}

\item{Hgiven_var}{Variable name for inheritance given}

\item{Hreceived_var}{Variable name for inheritance received}
}
\value{
A \link[data.table]{data.table} object
}
\description{
Create a series of moments for indirect inference, either using
 wealth survey or microsimulated population
}
\details{
\describe{
  \item{\code{create_moment}}{Either use wealth survey or microsimulated population}
  \item{create_moment_data}{Compute moments for both wealth survey and microsimulated population}
}

Order is the following:
\enumerate{
  \item Acme for wealth distribution by age (\link{moment_pic})
  \item Wealth evolution between two periods  (\link{moment_variation_K})
  \item Age where wealth is at its acme (\link{moment_age})
}
}
\seealso{
\link[capitulation]{life_cycle_model}
}
