% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_variation_K.R
\name{deltaK_simulations}
\alias{deltaK_simulations}
\title{Get wealth variation between too survey years in microsimulated data}
\usage{
deltaK_simulations(
  simulations,
  ages = c(35, 50),
  survey_years = c(2015, 2018),
  by = NULL,
  wealth_var = "wealth",
  scale = c("level", "log", "asinh"),
  exclude_negative = FALSE
)
}
\arguments{
\item{simulations}{Simulated longitudinal data}

\item{ages}{Ages conserved to compute moments}

\item{survey_years}{Survey data}

\item{wealth_var}{Wealth variable name}

\item{scale}{Scale for \code{wealth_var} that should
be applied}

\item{exclude_negative}{Should we exclude negative values
for wealth. Automatic if \code{scale} = \code{log}.}
}
\value{
Wide formatted wealth and age for survey_years for 
 people aged in ages brackets
}
\description{
Get wealth variation between too survey years in microsimulated data
}
