% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_price.R
\name{import_price}
\alias{import_price}
\title{Import price index used in \code{Destinie} model}
\usage{
import_price(path_destinie = "./inst/dataINSEE/Destinie")
}
\arguments{
\item{path_destinie}{Path to Destinie data}
}
\description{
Import price index used in \code{Destinie} model
}
