% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_survey.R
\name{longitudinal_survey}
\alias{longitudinal_survey}
\title{Create longitudinal wealth survey at household level}
\usage{
longitudinal_survey(macro, EP_2015, EP_2018, path_data)
}
\arguments{
\item{macro}{Macroeconomic table}

\item{EP_2015, EP_2018}{Wealth surveys (temporary
fix for the function to run)}

\item{path_data}{Directory where data are stored}
}
\value{
2015 and 2018 longitudinal wealth survey
}
\description{
Create longitudinal wealth survey at household level
}
