% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_capitulation.R
\name{model_capitulation}
\alias{model_capitulation}
\title{A wrapper to use capitulation package within GMM framework}
\usage{
model_capitulation(theta = c(r = 0.03, gamma = 0.5, beta = 0.92), ...)
}
\arguments{
\item{theta}{A named vector of structural parameters}

\item{...}{Additional arguments provided}
}
\description{
A wrapper to use capitulation package within GMM framework
}
