% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_age.R
\name{moment_age_multiple}
\alias{moment_age_multiple}
\title{Function to fit moment based on multiple ages}
\usage{
moment_age_multiple(
  EP_data,
  ages = c(20, 80),
  wealth_var = "PATFISOM",
  age_var = "AGEPR",
  moment_var = "moment_data",
  moments_before = 2,
  normalize = TRUE,
  scale = c("level", "log", "asinh"),
  survey_year = 2015
)
}
\arguments{
\item{EP_data}{Wealth survey data or microsimulated data}

\item{ages}{Thresholds (lower, upper) that should be considered
as sample}

\item{wealth_var}{Wealth variable name}

\item{age_var}{Age variable}

\item{moment_var}{Variable name}

\item{moments_before}{Number of moments computed before}

\item{normalize}{Logical value indicating whether we want
to normalize wealth variable to get variance equal to 1}

\item{scale}{Should values be considered in \emph{level} (default) or \emph{log}}

\item{survey_year}{Year that should be used to compute moments}
}
\description{
Function to fit moment based on multiple ages
}
