% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_pic.R
\name{moment_pic}
\alias{moment_pic}
\title{Acme of median wealth function by age (5 years cohort)}
\usage{
moment_pic(
  EP_data,
  wealth_var = "PATFISOM",
  age_var = "AGEPR",
  moment_var = "moment_data",
  normalize = FALSE,
  scale = c("log", "level", "asinh"),
  exclude_negative = FALSE,
  survey_year = 2015,
  moment_number = 1L,
  scale_moment1 = c("level", "share"),
  by = NULL
)
}
\arguments{
\item{EP_data}{Wealth survey}

\item{wealth_var}{Wealth variable in survey}

\item{age_var}{Age variable in survey}

\item{moment_var}{Variable name for moment result}

\item{normalize}{Logical value indicating whether we want to normalize
variable}

\item{scale}{Scale for the variable to consider. Can be level (default)
or log}

\item{exclude_negative}{Logical value indicating whether we should
exclude negative wealth values}

\item{survey_year}{Year that should be used to compute moments}

\item{moment_number}{Moment number that should be generated}
}
\value{
A \code{data.table} with 1 row with format
presented in \code{example} section
}
\description{
Acme of median wealth function by age (5 years cohort)
}
\examples{
\dontrun{
moment1 <- moment_pic(EP_2015)
moment1
# age_tranche       W
# 1:          14 47533.5
}
}
