% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moment_dK.R
\name{plot_moment_dK}
\alias{plot_moment_dK}
\title{Plot moment related to capital growth}
\usage{
plot_moment_dK(
  EP_lon,
  simulations,
  N_moments = 3L,
  scale = "log",
  statfunc = "median",
  deltaK_var = "dW",
  age_2015_var = "AGE_2015",
  ages = c(30, 65),
  normalize = FALSE,
  survey_years = c(2015, 2018),
  exclude_negative = FALSE,
  by = NULL,
  plot = TRUE,
  label_observed = "survey",
  label_simulated = "microsimulation"
)
}
\arguments{
\item{EP_lon}{Longitudinal data, either microsimulated or
wealth survey created by \link{longitudinal_survey}}

\item{simulations}{Microsimulated data}

\item{scale}{Should values be considered in \emph{level} (default) or \emph{log}}

\item{statfunc}{Statistical function (moment, quantile, etc. ) that should be used}

\item{deltaK_var}{Name for the growth variable that is temporarily created}

\item{age_2015_var}{Age variable storing age at first occurrence}

\item{ages}{Ages considered to restrict sample}

\item{normalize}{Logical value indicating whether values are normalized
to compare distributions with variance equal to 1}

\item{survey_years}{Collection years for the two survey waves}

\item{exclude_negative}{Should we exclude negative values? Default to \code{FALSE}}
}
\description{
Plot moment related to capital growth
}
