% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inheritance_sample.R
\name{prepare_inheritance_sample}
\alias{prepare_inheritance_sample}
\title{Prepare sample for inheritance analysis}
\usage{
prepare_inheritance_sample(
  path_survey = "../../../capitulation/inst/dataINSEE/Enquete Patrimoine",
  year = 2015
)
}
\arguments{
\item{path_survey}{Path to survey data}

\item{year}{Survey year}
}
\description{
Prepare sample for inheritance analysis
}
