% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_pic.R
\name{quantile_pic}
\alias{quantile_pic}
\title{Extract acme for wealth}
\usage{
quantile_pic(
  EP_data,
  wealth_var = "PATFISOM",
  age_var = "AGEPR",
  moment_var = "moment_data",
  normalize = TRUE,
  scale = c("level", "log", "asinh"),
  probs = seq(0.1, 0.9, by = 0.1),
  exclude_negative = FALSE,
  survey_year = 2015
)
}
\arguments{
\item{EP_data}{Wealth survey}

\item{wealth_var}{Wealth variable in survey}

\item{age_var}{Age variable in survey}

\item{moment_var}{Variable name for moment result}

\item{normalize}{Logical value indicating whether we want to normalize
variable}

\item{scale}{Scale for the variable to consider. Can be level (default)
or log}

\item{probs}{Probabilities to consider for quantile
computations}

\item{exclude_negative}{Logical value indicating whether we should
exclude negative wealth values}

\item{survey_year}{Year that should be used to compute moments}
}
\description{
Extract acme for wealth
}
