% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_variation_K.R
\name{quantile_variation_K}
\alias{quantile_variation_K}
\title{Get wealth variation between too survey years in microsimulated data}
\usage{
quantile_variation_K(
  EP_lon,
  format = "wide",
  wealth_var = "wealth",
  deltaK_var = "dW",
  moment_var = "moment_data",
  age_2015_var = "AGEPR_2015",
  ages = c(35, 50),
  survey_years = c(2015, 2018),
  stat = "proportion",
  normalize = TRUE,
  scale = c("level", "log", "asinh"),
  probs = seq(0.1, 0.9, by = 0.1),
  exclude_negative = FALSE
)
}
\arguments{
\item{EP_lon}{Longitudinal data, either microsimulated or
wealth survey created by \link{longitudinal_survey}}

\item{format}{Data structure. Either \emph{wide} (default)
or \emph{long}}

\item{wealth_var}{Wealth variable}

\item{deltaK_var}{Name for the growth variable that is temporarily created}

\item{moment_var}{Name for the moment variable that is created}

\item{age_2015_var}{Age variable storing age at first occurrence}

\item{ages}{Ages considered to restrict sample}

\item{survey_years}{Collection years for the two survey waves}

\item{stat}{Should we collect $W_{1}/W_{0}$ (\emph{proportion}, default)
or $W_{1}-W_{0}$ (\emph{difference})}

\item{normalize}{Logical value indicating whether values are normalized
to compare distributions with variance equal to 1}

\item{scale}{Should values be considered in \emph{level} (default) or \emph{log}}

\item{probs}{Probabilities to consider for quantile computations}

\item{exclude_negative}{Should we exclude negative values? Default to \code{FALSE}}
}
\description{
Get wealth variation between too survey years in microsimulated data
}
