% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EP.R
\name{read_EP}
\alias{read_EP}
\title{Read wealth survey to prepare moment computations}
\usage{
read_EP(
  macro,
  path_data = getwd(),
  level = c("household", "individual"),
  year = 2015,
  .colsWealth = NULL
)
}
\arguments{
\item{macro}{Macroeconomic table}

\item{path_data}{Directory where data are stored}

\item{level}{Analysis level. Default to 'household'}

\item{year}{Survey year}

\item{.colsWealth}{If needed, columns to keep. If
\code{NULL}, alls columns are kept}
}
\value{
Wealth survey
}
\description{
Read wealth survey to prepare moment computations
}
