% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EP.R
\name{read_wealth}
\alias{read_wealth}
\title{Import wealth survey data}
\usage{
read_wealth(
  path = "./inst/dataINSEE/Enquete Patrimoine",
  tabname = "basecomplete0_62663.sas7bdat",
  create_education = FALSE
)
}
\arguments{
\item{path}{Path for wealth survey}

\item{tabname}{SAS table name}

\item{create_education}{Logical value indicating whether we
should create an education variable}
}
\description{
Import wealth survey data
}
