% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_distribution_K.R
\name{stat_distribution_K}
\alias{stat_distribution_K}
\alias{moment_distribution_K}
\title{Cross sectional moments for wealth distribution}
\usage{
stat_distribution_K(
  EP_data,
  quantiles = c(0.1, 0.5, 0.9),
  wealth_var = "PATFISOM",
  moment_var = "moment_data",
  normalize = TRUE,
  scale = c("log", "level"),
  exclude_negative = TRUE,
  survey_year = 2015,
  min_moment = 1L
)

moment_distribution_K(
  EP_data,
  stats = c("mean", "sd"),
  wealth_var = "PATFISOM",
  moment_var = "moment_data",
  normalize = TRUE,
  scale = c("log", "level"),
  exclude_negative = TRUE,
  survey_year = 2015,
  min_moment = 1L
)
}
\arguments{
\item{EP_data}{Wealth survey}

\item{quantiles}{Quantiles that should be computed}

\item{wealth_var}{Wealth variable in survey}

\item{moment_var}{Variable name for moment result}

\item{normalize}{Logical value indicating whether we want to normalize
variable}

\item{scale}{Scale for the variable to consider. Can be level (default)
or log}

\item{exclude_negative}{Logical value indicating whether we should
exclude negative wealth values}

\item{survey_year}{Year that should be used to compute moments}

\item{min_moment}{Minimal number for moment identifier}

\item{stats}{Moment that should be used. This must be a
data summary statistics that should match a base R function}
}
\value{
A \link[data.table]{data.table} with three columns:
 \code{moment_var} (function argument), \code{p} (quantile considered)
 and \code{Nmoment}
}
\description{
Compute population or simulated moments
 for wealth distribution
}
